end Crud laravel

+++ app/Http/Controllers/EmpleadoController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Empleado;
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Storage;
+
 class EmpleadoController extends Controller
 {
     /**
@@ -15,7 +17,7 @@ class EmpleadoController extends Controller
     public function index()
     {
         //
-        $datos['empleados']=Empleado::paginate(0);
+        $datos['empleados']=Empleado::paginate(1);
         return view('empleado.index',$datos );
     }
 
@@ -40,6 +42,22 @@ class EmpleadoController extends Controller
     {
         //
         // $datosEmpleado = request()->all();
+        
+        $campos=[
+            'Nombre'=>'required|string|max:25',
+            'ApellidoPaterno'=>'required|string|max:30',
+            'ApellidoMaterno'=>'required|string|max:30',
+            'Correo'=>'required|email',
+            'Foto'=>'required|max:10000|mimes:jpeg,png,jpg',  
+        ];
+        $mensaje=[
+            'required'=>'El :attribute es requerido',
+            'Foto.required'=>'La foto es requerida'
+        
+        ];
+        $this->validate($request, $campos,$mensaje);
+
+
         $datosEmpleado = request()->except('_token');
         
         if($request->hasFile('Foto')){
@@ -47,7 +65,8 @@ class EmpleadoController extends Controller
         }
         
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('/empleado')->with('mensaje','Empleado agregado con exito');
     }
 
     /**
@@ -82,9 +101,43 @@ class EmpleadoController extends Controller
      * @param  \App\Models\Empleado  $empleado
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Empleado $empleado)
+    public function update(Request $request, $id)
+
     {
         //
+        $campos=[
+            'Nombre'=>'required|string|max:25',
+            'ApellidoPaterno'=>'required|string|max:30',
+            'ApellidoMaterno'=>'required|string|max:30',
+            'Correo'=>'required|email',
+        ];
+        $mensaje=['required'=>'El :attribute es requerido',];
+        
+        if($request->hasFile('Foto')){
+            $campos=['Foto'=>'required|max:10000|mimes:jpeg,png,jpg',];  
+            $mensaje=['Foto.required'=>'La foto es requerida'];
+
+        }
+
+        $this->validate($request, $campos,$mensaje);
+
+
+
+        $datosEmpleado = request()->except(['_token', '_method']);
+
+        if($request->hasFile('Foto')){
+            $empleado=Empleado::findOrFail($id);
+            
+            Storage::delete('public/'.$empleado->Foto);
+
+
+            $datosEmpleado['Foto']=$request->file('Foto')->store('uploads', 'public');
+            }
+
+        Empleado::where('id', '=',$id)->update($datosEmpleado);
+        $empleado=Empleado::findOrFail($id);
+        //return view('empleado.edit', compact('empleado'));
+        return redirect('empleado')->with('mensaje','Empleado actualizado');
     }
 
     /**
@@ -96,7 +149,15 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('/empleado');
+        
+        $empleado=Empleado::findOrFail($id);
+
+        if(Storage::delete('public/'.$empleado->Foto)){
+            
+            Empleado::destroy($id);
+
+        }
+
+        return redirect('/empleado')->with('mensaje','Empleado eliminado');
     }
 }
